แก้บัค bread.cost เป็น doubleแล้ว ในsummary และ Myform

+++ Project1/bread.cpp
@@ -1,5 +1,5 @@
 #include "bread.h"
 
-int bread::amount(int sold, int cost) {
+double bread::amount(int sold, double cost) {
     return sold * cost;
 }
